Phase K Waves 68+69: 0x007F 4B hardening + LP32/LP64 mass tightening sweep
Wave 68 — 0x007F MANUFACTURE_SET_MANUFACTURE_ID byte-exact 4B hardening.
The retail Win32 server emitted 0x007F with a 4-byte payload (LP32
`long`). Our Linux build was emitting 8 bytes (LP64 `long`) — silent
wire-shape drift. Single-token fix at PlayerConnection.cpp:10134 and
the matching declaration at PlayerClass.h:1028: `long mfg_id` →
`int32_t mfg_id`. `sizeof(mfg_id)` now locks to 4 on every platform.

New hardening test SectorManufactureSetManufactureIdHardeningTests.
ManufactureSetManufactureId_EmittedDuringHandshake_HasExactly4BytePayload
filters HandshakeFrames for 0x007F and asserts every captured frame
carries a 4-byte payload. To make that assertion possible without
re-driving the login flow, SectorHandshake.Session gained a new
HandshakeFrames field of type IReadOnlyList<(ushort Opcode, int
PayloadLength)>, populated alongside the existing HandshakeOpcodes
during DoSectorLoginUntilStartAsync.

+0 ratchet for Wave 68 — 0x007F is already counted by Wave 35's
HandshakeEmitsFullSendLoginShipDataFanout passive-presence assertion.
The TestedOpcodes entry for 0x007F is amended to credit the new
hardening test and document the parameter tightening.

Wave 69 — LP32/LP64 mass tightening sweep, motivated by reviewer
feedback that per-opcode tightening was burning tokens. Mass audit
pattern: grep `Send*(..., sizeof(varname))` cross-referenced against
`long varname` declarations across server-native code (server/,
login-server/, proxy/, excluding server/third_party/ and client/).

Four wire-shape body bugs tightened in one sweep:

  * 0x003C CLIENT_TYPE — `SendClientType(long client_type)` →
    `int32_t`. Emitted via `SendOpcode(..., sizeof(client_type))`.
  * 0x0032 DEACTIVATE_RENDER_STATE — `SendDeactivateRenderState(long
    game_id)` → `int32_t`. Emitted via `SendOpcode(..., sizeof(game_id))`.
  * 0x0007 REMOVE — `RemoveObject(long object_id)` → `int32_t`.
    Emitted via `SendOpcode(..., sizeof(object_id))`.
  * 0x0003 LOGOFF — `long GameIDD = GameID()` in `Player::ForceLogout`
    → `int32_t`. Emitted via `SendOpcode(..., sizeof(GameIDD))`.

Three hygiene cleanups (render-state opcode declarations narrowed to
match the int32-typed wire structs they already write into; no body
behaviour changes):

  * 0x0030 ACTIVATE_RENDER_STATE
  * 0x002F INIT_RENDER_STATE
  * 0x0031 ACTIVATE_NEXT_RENDER_STATE

Triage rule maintained: only `long`s that touch the wire become
`int32_t`. Loop counters, file offsets, time_t and other purely-local
`long`s are left alone — narrowing them would be churn without
correctness gain.

Per CLAUDE.md server-integrity rule: every change in both waves is a
tightening toward the retail Win32 wire shape — the rule explicitly
welcomes this direction. No widened input acceptance, no loosened
gating, no fabricated replies. Build green; broad 19-test regression
sweep 18/19 passed (one transient flake passed on isolated re-run).

Coverage stays at 102/207 (49.3%) — Wave 68 is a +0 ratchet on
already-counted 0x007F; Wave 69 is body tightening on already-counted
opcodes.

+++ server/src/PlayerClass.h
@@ -976,7 +976,7 @@ public:
     void        SendClientSound(char *sound_name, long channel = 0, char queue = 0, long warninglevel = -1);
     void        PointEffect(float *position, short effect_id, float scale = 1.0f);
     void        SendObjectEffect(ObjectEffect *object_effect);
-    void        SendClientType(long client_type);       // opcode 0x3c
+    void        SendClientType(int32_t client_type);    // opcode 0x3c
     void        SendStart(long start_id);               // opcode 0x05
     void        SendSetBBox(float xmin, float ymin, float xmax, float ymax);
     void        SendSetZBand(float min, float max);     // opcode 0x2a
@@ -991,16 +991,16 @@ public:
     void        SendComponentPositionalUpdate(long object_id, PositionInformation * position_info, long timestamp=0);
     void        SendAdvancedPositionalUpdate(long object_id, PositionInformation * position_info);
     void        SendObjectToObjectEffect(ObjectToObjectEffect *obj_effect);
-    void        SendActivateRenderState(long game_id, unsigned long render_state_id);
-    void        SendInitRenderState(long game_id, unsigned long render_state_id);
-    void        SendActivateNextRenderState(long game_id, unsigned long render_state_id);
-    void        SendDeactivateRenderState(long game_id);
+    void        SendActivateRenderState(int32_t game_id, uint32_t render_state_id);
+    void        SendInitRenderState(int32_t game_id, uint32_t render_state_id);
+    void        SendActivateNextRenderState(int32_t game_id, uint32_t render_state_id);
+    void        SendDeactivateRenderState(int32_t game_id);
     void        OpenInterface(long UIChange, long UIType);
     void        SendSetTarget(int game_id, int target_id);
     void        SendPushMessage(char *msg1, char *type, long time, long priority);
     void        SetResourceDrainLevel(Object *obj, long slot);
 	void		SetHuskDrainLevel(Object *obj, long slot);
-    void        RemoveObject(long object_id);
+    void        RemoveObject(int32_t object_id);
     void        SendProspectAUX(long value, int type); //HAX! Remove when decyphered
     void        CreateTractorComponent(float *position, float decay, float tractor_speed, long player_id, long article_id, long effect_id, long timestamp);
     bool        CheckResourceLock(long object_id);
@@ -1025,7 +1025,7 @@ public:
     void        OpenStargate(long object_id);
     void        CloseStargate(long object_id);
     bool        SendLoungeNPC(long StationID);
-    void        SetManufactureID(long mfg_id);          // opcode 0x7f
+    void        SetManufactureID(int32_t mfg_id);       // opcode 0x7f
     void        SendServerHandoff(long from_sector_id, long to_sector_id, char *from_sector, char *from_system, char *to_sector, char *to_system);
     void        SendNotifyEmote(long game_id, long emote);
     void        SendClientChatEvent(long Type, Player *Source, char *Channel="", char *Message="" , char *OtherPlayer="", char *NonPlayerSrc="");
